Split storage.py
