+++ torchrl/__init__.py
@@ -1,2 +1,1 @@
-from torchrl.storage import ReplayBuffer, CPUReplayBuffer
 from torchrl.episode_runner import MultiEpisodeRunner
